Fix #911: Convert table name for manual relationship (remote_table) -  object or a string (in case of public schema)
