+++ console/src/components/Services/Data/TablePermissions/PermissionBuilder/PermissionBuilder.js
@@ -15,6 +15,8 @@ import {
   boolColumnOperators,
 } from './utils';
 
+import { getTableName } from '../../utils';
+
 import QueryBuilderJson from '../../../../QueryBuilderJson/QueryBuilderJson';
 
 const styles = require('./Styles.scss');
@@ -406,7 +408,8 @@ class PermissionBuilder extends React.Component {
       let tableColumns = [];
       let tableRelationships = [];
       if (table) {
-        const tableSchema = tableSchemas[table];
+        // In case of a manual relationship the right table is an object for a different schema
+        const tableSchema = tableSchemas[getTableName(table)];
         tableColumns = getTableColumnNames(tableSchema);
         tableRelationships = getTableRelationshipNames(tableSchema);
       }
