+++ console/src/components/Services/Data/utils.js
@@ -119,6 +119,16 @@ const escapeRegExp = string => {
   return string.replace(/([.*+?^${}()|[\]\\])/g, '\\$1');
 };
 
+const getTableName = t => {
+  const typ = typeof t;
+  if (typ === 'string') {
+    return t;
+  } else if (typ === 'object') {
+    return 'name' in t ? t.name : '';
+  }
+  return '';
+};
+
 export {
   ordinalColSort,
   findTableFromRel,
@@ -126,4 +136,5 @@ export {
   getEdForm,
   getIngForm,
   escapeRegExp,
+  getTableName,
 };
